# specs/001-property-sales-platform/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Property Sales Platform API
  description: RESTful API for property browsing, listing management, AI-assisted inquiries, and user favorites
  version: 1.0.0
  contact:
    name: Property Sales Platform
    email: support@propertyplatform.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://propertyplatform.vercel.app/api
    description: Production server
...
paths:
...
  /properties/{id}:
    get:
      tags: [Properties]
      summary: Get property details
      description: Get detailed information about a specific property
      operationId: getPropertyById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Property ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    
    put:
      tags: [Properties]
      summary: Update property listing
      description: Update an existing property listing (owner only)
      operationId: updateProperty
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertyRequest'
      responses:
        '200':
          description: Property updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    
    delete:
      tags: [Properties]
      summary: Delete property listing
      description: Delete a property listing (owner only)
      operationId: deleteProperty
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Property deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    PropertyDetail:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            photos:
              type: array
              items:
                $ref: '#/components/schemas/PropertyPhoto'
            owner:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
            representative:
              type: object
              properties:
                name:
                  type: string
            isFavorited:
              type: boolean
    Property:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        propertyType:
          type: string
          enum: [HOUSE, APARTMENT, CONDO, TOWNHOUSE]
        price:
          type: number
        bedrooms:
          type: integer
        bathrooms:
          type: number
        squareFeet:
          type: integer
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ACTIVE, SOLD, PENDING, ARCHIVED]
        soldDate:
          type: string
          format: date-time
          nullable: true
        ownerId:
          type: string
        representativeId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdatePropertyRequest:
      type: object
      properties:
        title:
          type: string
        price:
          type: number
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ACTIVE, SOLD, PENDING, ARCHIVED]
    PropertyPhoto:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        displayOrder:
          type: integer
        uploadedAt:
          type: string
          format: date-time
    
    # Inquiry Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'